Added sample personal rules
Personal rules is a financial format that is provided to the money tool to transform a list of transactions into personalized reports.

The format uses JSON to describe a simple object structure.

## Structure:

### monthlyIncome
The monthly income represents a fixed amount of money that you will take in per month. This number is used to estimate your future financial state. This value will not be used to represent actual account balances.

### categories
The categories property accepts an array of category rules. These category rules do not specify which categories relate to which transactions. That is handled in the transaction section. These category rules are overrides. An category override is a rule that is applied after a category value is already assigned to a transaction. For example, let's say you want to automatically flag certain fast food transactions.

Using an override, you can replace the category of any "Fast Food" category with the new category "Fast Food (Tom)"" when transaction has the category "Fast Food" and the transaction name contains the word "Naperville". This example is relevant, because I want to track all my fast food transactions. These transactions only happen when I am a few towns away from home, so they are easy to identify.

### transactions
The transactions property accepts an array of transaction rules. This is the most common transformation you can make against your financial data. These rules allow you to customize the name, category and need/want.

Acceptable fields:
- contains (string|[]string)
This searches the transaction name to see if it matches a string or array of strings.

- replace (string)
If a match is found, the transaction name is replaced.

- category (string)
Assigns the category to the string if a match is found.

- need (boolean)
Assigns whether the transaction is a need/want

- findReplace (string)
This searches the string for the provided string and replaces the entire transaction name with the string. This is useful if you want to find all Chipotle transactions and replace the garbled transaction with Chipotle.

- bills
This field will be described in a future update.

## FAQ
1. What is this need/want feature?
This is a unique category that is applied to all transactions to label whether the transaction is important or not. If you aren't reaching your savings goals, you can use this feature to identify wants. Ideally, you will make smarter financial decisions by removing want transactions.

2. How will savings work?
I am still working this out.

3. How will bills work?
Bills will be defined per transaction rule. For example, if you have an electric bill, you will be able to specify the expected date of the transaction. You will also be able to specify the expected range that the bill should appear. Typically, bills will be fixed, but some are variable.

4. How will allowances work?
Allowances will be calculated based on your total income - bills - savings