Updated ReadMe
I don’t like the structure of the account system.

One intent of this program is simplicity. I don’t want to see accounts broken out. An account is a detail.

The only thing that matters on an account is the starting balance.

I will need to modify the program to work as only a list of transactions. I will than need to remove much of the logic around accounts and figure out how to get the starting balance.