Bills are now mapped to the transaction
Bills will become more valuable once Reports are available.

For example, we can build a report that shows a list of all known bills with the dates the bills appear. We can than check all of the transactions to see if the bill has been paid. If it has, than we can show that bill is no longer expecting.