Added Bill Report
Bills displays a list of all bills provided by the rules.
Each list item displays the what was actually spent on the bill
and how much should have been spent on the bill.

The purpose of this view is to provide a simple summary of what bills
have been handled and which ones still need to be handled.

Another important value in this view is the projected bill spend.
The projected bill spend is the total amount spent on bills so far based
on transaction history plus the total amount remaining on bills. This value
can change over the course of the month. It is possible to have a lower projected
goal if you spend less on your bills than expected. It is pssible to have a higher
projected goal if you spend more on a bill. Variable bills such as an Electric
or Gas bill are a good reason why the projected value can change.